CLIM-1367: Reusable ModalSummaryPopover
